backend/requirements.txt
# Backend Dependencies

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Database (Future)
# sqlalchemy>=2.0.0
# alembic>=1.13.0

# AI/ML (Future)
# openai>=1.0.0
# langchain>=0.1.0
